+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/livesmoke/LiveSmokeNavigationFlows.kt
@@ -1,0 +1,61 @@
+@file:OptIn(androidx.compose.ui.test.ExperimentalTestApi::class)
+
+package com.flashcardsopensourceapp.app.livesmoke
+
+import androidx.compose.ui.test.hasClickAction
+import androidx.compose.ui.test.hasSetTextAction
+import androidx.compose.ui.test.hasText
+import androidx.compose.ui.test.onAllNodesWithText
+import androidx.compose.ui.test.performTextReplacement
+
+internal fun LiveSmokeContext.openCardsTab() {
+    clickNode(
+        matcher = hasText("Cards").and(other = hasClickAction()),
+        label = "Cards tab"
+    )
+}
+
+internal fun LiveSmokeContext.openReviewTab() {
+    clickNode(
+        matcher = hasText("Review").and(other = hasClickAction()),
+        label = "Review tab"
+    )
+}
+
+internal fun LiveSmokeContext.openAiTab() {
+    clickNode(
+        matcher = hasText("AI").and(other = hasClickAction()),
+        label = "AI tab"
+    )
+}
+
+internal fun LiveSmokeContext.openSettingsTab() {
+    clickNode(
+        matcher = hasText("Settings").and(other = hasClickAction()),
+        label = "Settings tab"
+    )
+}
+
+internal fun LiveSmokeContext.openSettingsSection(sectionTitle: String) {
+    openSettingsTab()
+    waitUntilAtLeastOneExistsOrFail(
+        matcher = hasText(sectionTitle),
+        timeoutMillis = internalUiTimeoutMillis
+    )
+    clickNode(
+        matcher = hasText(sectionTitle).and(other = hasClickAction()),
+        label = sectionTitle
+    )
+}
+
+internal fun LiveSmokeContext.dismissAiConsentIfNeeded() {
+    if (composeRule.onAllNodesWithText("Before you use AI").fetchSemanticsNodes().isNotEmpty()) {
+        clickText(text = "OK", substring = false)
+    }
+}
+
+internal fun LiveSmokeContext.updateCardText(fieldTitle: String, value: String) {
+    clickText(text = fieldTitle, substring = false)
+    composeRule.onAllNodes(hasSetTextAction())[0].performTextReplacement(value)
+    tapBackIcon()
+}
