+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/livesmoke/LiveSmokeWorkspaceLifecycleFlows.kt
@@ -2,47 +2,19 @@
 
 package com.flashcardsopensourceapp.app.livesmoke
 
-import androidx.compose.ui.semantics.SemanticsProperties
-import androidx.compose.ui.semantics.getOrNull
-import androidx.compose.ui.test.hasClickAction
-import androidx.compose.ui.test.hasSetTextAction
 import androidx.compose.ui.test.hasTestTag
 import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.onAllNodesWithTag
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollToNode
-import androidx.compose.ui.test.performTextInput
 import androidx.compose.ui.test.performTextReplacement
 import com.flashcardsopensourceapp.app.FlashcardsApplication
 import com.flashcardsopensourceapp.data.local.model.CloudAccountState
-import com.flashcardsopensourceapp.feature.settings.cloud.cloudPostAuthWorkspaceRowTag
-import com.flashcardsopensourceapp.feature.settings.cloud.cloudPostAuthExistingButtonTag
-import com.flashcardsopensourceapp.feature.settings.cloud.cloudSignInEmailFieldTag
-import com.flashcardsopensourceapp.feature.settings.cloud.cloudSignInSendCodeButtonTag
 import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceCreateButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceErrorMessageTag
 import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceExistingRowTag
 import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceListTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceLoadingStateTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceNameTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceOperationMessageTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceReloadButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceSelectedSummaryTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressErrorTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressConfirmationButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressConfirmationDialogTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressConfirmationFieldTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressConfirmationPhraseTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressPreviewBodyTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressPreviewButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressPreviewDialogTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewTodayDueCountTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewTodayNewCountTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewTodayReviewedCountTag
 import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteConfirmationButtonTag
 import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteConfirmationDialogTag
 import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteConfirmationErrorTag
@@ -53,7 +25,6 @@ import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewD
 import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeletePreviewContinueButtonTag
 import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeletePreviewDialogTag
 import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteWorkspaceButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewErrorMessageTag
 import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewNameFieldTag
 import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewSaveNameButtonTag
 import kotlinx.coroutines.flow.combine
@@ -77,117 +48,6 @@ internal data class EphemeralWorkspaceHandle(
     val workspaceId: String,
     val workspaceName: String
 )
-
-private const val linkedSignInTimeoutMillis: Long = 120_000L
-private const val cloudPostAuthLinkingWorkspaceTitle: String = "Linking workspace"
-private const val cloudPostAuthSyncingWorkspaceTitle: String = "Syncing workspace"
-private const val cloudPostAuthGuestUpgradeTitle: String = "Upgrading guest account"
-private const val cloudPostAuthRetryButtonText: String = "Retry"
-private const val accountStatusSyncNowButtonText: String = "Sync now"
-
-internal fun LiveSmokeContext.signInWithReviewAccount(reviewEmail: String) {
-    openSettingsTab()
-    clickNode(
-        matcher = hasText("Account").and(other = hasClickAction()),
-        label = "Account"
-    )
-    clickText(text = "Account status", substring = false)
-    clickText(text = "Sign in or sign up", substring = false)
-    composeRule.onNodeWithTag(cloudSignInEmailFieldTag).performTextInput(reviewEmail)
-    clickTag(tag = cloudSignInSendCodeButtonTag, label = "Send code")
-
-    waitForCloudSignInSurface()
-    completeCloudPostAuthWorkspaceSelectionIfNeeded()
-    waitForLinkedAccountStatusAfterSignIn()
-    tapBackIcon()
-    tapBackIcon()
-}
-
-private fun LiveSmokeContext.waitForCloudSignInSurface() {
-    try {
-        waitUntilWithMitigation(
-            timeoutMillis = linkedSignInTimeoutMillis,
-            context = "while waiting for sign-in to reach account status or post-auth sync"
-        ) {
-            hasVisibleText(text = accountStatusSyncNowButtonText, substring = false) ||
-                hasVisibleText(text = cloudSyncChooserPrompt, substring = false) ||
-                hasVisibleText(text = cloudPostAuthLinkingWorkspaceTitle, substring = false) ||
-                hasVisibleText(text = cloudPostAuthSyncingWorkspaceTitle, substring = false) ||
-                hasVisibleText(text = cloudPostAuthGuestUpgradeTitle, substring = false) ||
-                hasVisibleText(text = cloudPostAuthRetryButtonText, substring = false)
-        }
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Sign-in did not reach account status or post-auth sync. " +
-                "CloudSettings=${currentCloudSettingsSummary()} " +
-                "CurrentWorkspace=${currentWorkspaceSummaryOrNull()} " +
-                "VisibleRows=${captureVisibleWorkspaceRows(rowTag = cloudPostAuthWorkspaceRowTag)} " +
-                "SystemDialog=${currentBlockingSystemDialogSummaryOrNull()}",
-            error
-        )
-    }
-}
-
-private fun LiveSmokeContext.completeCloudPostAuthWorkspaceSelectionIfNeeded() {
-    if (hasVisibleText(text = cloudSyncChooserPrompt, substring = false).not()) {
-        return
-    }
-
-    val visibleRows: List<String> = captureVisibleWorkspaceRows(rowTag = cloudPostAuthWorkspaceRowTag)
-    if (visibleRows.isEmpty()) {
-        throw AssertionError(
-            "Cloud sync chooser was visible without selectable workspace rows. " +
-                "CloudSettings=${currentCloudSettingsSummary()} " +
-                "CurrentWorkspace=${currentWorkspaceSummaryOrNull()}"
-        )
-    }
-
-    runWithInlineRawScreenStateOnFailure(action = "click_post_auth_workspace_row") {
-        dismissExternalSystemDialogIfPresent()
-        val preferredWorkspaceId: String? = currentActiveWorkspaceIdOrNull()
-        if (
-            preferredWorkspaceId != null &&
-            composeRule.onAllNodesWithTag(
-                testTag = cloudPostAuthExistingButtonTag(workspaceId = preferredWorkspaceId)
-            ).fetchSemanticsNodes().isNotEmpty()
-        ) {
-            composeRule.onNodeWithTag(
-                testTag = cloudPostAuthExistingButtonTag(workspaceId = preferredWorkspaceId)
-            ).performClick()
-        } else {
-            composeRule.onAllNodesWithTag(cloudPostAuthWorkspaceRowTag)[0].performClick()
-        }
-        composeRule.waitForIdle()
-    }
-}
-
-private fun LiveSmokeContext.waitForLinkedAccountStatusAfterSignIn() {
-    try {
-        waitUntilWithMitigation(
-            timeoutMillis = linkedSignInTimeoutMillis,
-            context = "while waiting for the linked account status surface after sign-in"
-        ) {
-            if (hasVisibleText(text = cloudPostAuthRetryButtonText, substring = false)) {
-                throw AssertionError(
-                    "Cloud post-auth failed after sign-in. " +
-                        "CloudSettings=${currentCloudSettingsSummary()} " +
-                        "CurrentWorkspace=${currentWorkspaceSummaryOrNull()} " +
-                        "VisibleRows=${captureVisibleWorkspaceRows(rowTag = cloudPostAuthWorkspaceRowTag)}"
-                )
-            }
-            hasVisibleText(text = accountStatusSyncNowButtonText, substring = false)
-        }
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Linked account status did not appear after sign-in. " +
-                "CloudSettings=${currentCloudSettingsSummary()} " +
-                "CurrentWorkspace=${currentWorkspaceSummaryOrNull()} " +
-                "VisibleRows=${captureVisibleWorkspaceRows(rowTag = cloudPostAuthWorkspaceRowTag)} " +
-                "SystemDialog=${currentBlockingSystemDialogSummaryOrNull()}",
-            error
-        )
-    }
-}
 
 internal fun LiveSmokeContext.createEphemeralWorkspace(workspaceName: String): EphemeralWorkspaceHandle {
     openSettingsTab()
@@ -500,320 +360,6 @@ private fun LiveSmokeContext.tapDeleteWorkspaceConfirmation(workspaceName: Strin
     }
 }
 
-internal fun LiveSmokeContext.openCardsTab() {
-    clickNode(
-        matcher = hasText("Cards").and(other = hasClickAction()),
-        label = "Cards tab"
-    )
-}
-
-internal fun LiveSmokeContext.openReviewTab() {
-    clickNode(
-        matcher = hasText("Review").and(other = hasClickAction()),
-        label = "Review tab"
-    )
-}
-
-internal fun LiveSmokeContext.openAiTab() {
-    clickNode(
-        matcher = hasText("AI").and(other = hasClickAction()),
-        label = "AI tab"
-    )
-}
-
-internal fun LiveSmokeContext.openSettingsTab() {
-    clickNode(
-        matcher = hasText("Settings").and(other = hasClickAction()),
-        label = "Settings tab"
-    )
-}
-
-internal fun LiveSmokeContext.openSettingsSection(sectionTitle: String) {
-    openSettingsTab()
-    waitUntilAtLeastOneExistsOrFail(
-        matcher = hasText(sectionTitle),
-        timeoutMillis = internalUiTimeoutMillis
-    )
-    clickNode(
-        matcher = hasText(sectionTitle).and(other = hasClickAction()),
-        label = sectionTitle
-    )
-}
-
-internal fun LiveSmokeContext.resetWorkspaceProgressFromSettings(expectedCardsToResetCount: Int) {
-    openSettingsSection(sectionTitle = "Workspace")
-    clickTag(tag = workspaceSettingsResetProgressButtonTag, label = "Reset all progress")
-    waitForResetProgressConfirmationReady()
-
-    val confirmationPhrase = readTaggedText(
-        tag = workspaceSettingsResetProgressConfirmationPhraseTag,
-        label = "reset progress confirmation phrase"
-    )
-    composeRule.onNodeWithTag(workspaceSettingsResetProgressConfirmationFieldTag)
-        .performTextReplacement(confirmationPhrase)
-    clickTag(
-        tag = workspaceSettingsResetProgressConfirmationButtonTag,
-        label = "Continue reset progress"
-    )
-
-    waitForResetProgressPreviewReady(expectedCardsToResetCount = expectedCardsToResetCount)
-    clickTag(tag = workspaceSettingsResetProgressPreviewButtonTag, label = "Confirm reset progress")
-    waitForTagToDisappear(
-        tag = workspaceSettingsResetProgressPreviewDialogTag,
-        timeoutMillis = externalUiTimeoutMillis,
-        context = "while waiting for reset progress to complete"
-    )
-}
-
-internal fun LiveSmokeContext.assertWorkspaceTodayCounts(
-    expectedDueCount: Int,
-    expectedNewCount: Int,
-    expectedReviewedCount: Int
-) {
-    openSettingsSection(sectionTitle = "Workspace")
-    clickNode(
-        matcher = hasText("Overview").and(other = hasClickAction()),
-        label = "Workspace overview"
-    )
-    waitUntilWithMitigation(
-        timeoutMillis = externalUiTimeoutMillis,
-        context = "while waiting for workspace overview counts"
-    ) {
-        readTaggedTextOrNull(workspaceOverviewTodayDueCountTag) != null &&
-            readTaggedTextOrNull(workspaceOverviewTodayNewCountTag) != null &&
-            readTaggedTextOrNull(workspaceOverviewTodayReviewedCountTag) != null
-    }
-
-    val dueCount = readTaggedText(
-        tag = workspaceOverviewTodayDueCountTag,
-        label = "workspace overview due count"
-    ).toInt()
-    val newCount = readTaggedText(
-        tag = workspaceOverviewTodayNewCountTag,
-        label = "workspace overview new count"
-    ).toInt()
-    val reviewedCount = readTaggedText(
-        tag = workspaceOverviewTodayReviewedCountTag,
-        label = "workspace overview reviewed count"
-    ).toInt()
-
-    if (
-        dueCount != expectedDueCount ||
-        newCount != expectedNewCount ||
-        reviewedCount != expectedReviewedCount
-    ) {
-        throw AssertionError(
-            "Workspace overview counts did not match. " +
-                "Expected due/new/reviewed=$expectedDueCount/$expectedNewCount/$expectedReviewedCount " +
-                "but was $dueCount/$newCount/$reviewedCount. " +
-                "CloudSettings=${currentCloudSettingsSummary()} " +
-                "CurrentWorkspace=${currentWorkspaceSummaryOrNull()}"
-        )
-    }
-}
-
-private fun LiveSmokeContext.waitForResetProgressConfirmationReady() {
-    try {
-        waitForTagToExist(
-            tag = workspaceSettingsResetProgressConfirmationDialogTag,
-            timeoutMillis = externalUiTimeoutMillis,
-            context = "while waiting for reset progress confirmation dialog"
-        )
-        waitUntilWithMitigation(
-            timeoutMillis = externalUiTimeoutMillis,
-            context = "while waiting for reset progress confirmation fields"
-        ) {
-            composeRule.onAllNodesWithTag(workspaceSettingsResetProgressConfirmationPhraseTag)
-                .fetchSemanticsNodes()
-                .isNotEmpty() &&
-                composeRule.onAllNodesWithTag(workspaceSettingsResetProgressConfirmationFieldTag)
-                    .fetchSemanticsNodes()
-                    .isNotEmpty()
-        }
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Reset progress confirmation dialog did not become ready. " +
-                "ConfirmationPhrase=${readTaggedTextOrNull(workspaceSettingsResetProgressConfirmationPhraseTag)} " +
-                "ConfirmationError=${workspaceSettingsResetProgressErrorTextOrNull()} " +
-                "CloudSettings=${currentCloudSettingsSummary()} " +
-                "CurrentWorkspace=${currentWorkspaceSummaryOrNull()}",
-            error
-        )
-    }
-}
-
-private fun LiveSmokeContext.waitForResetProgressPreviewReady(expectedCardsToResetCount: Int) {
-    try {
-        waitForTagToExist(
-            tag = workspaceSettingsResetProgressPreviewDialogTag,
-            timeoutMillis = externalUiTimeoutMillis,
-            context = "while waiting for reset progress preview dialog"
-        )
-        waitUntilWithMitigation(
-            timeoutMillis = externalUiTimeoutMillis,
-            context = "while waiting for reset progress preview body"
-        ) {
-            val previewBodyText = readTaggedTextOrNull(workspaceSettingsResetProgressPreviewBodyTag)
-            previewBodyText != null &&
-                extractResetProgressCount(previewBodyText) == expectedCardsToResetCount
-        }
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Reset progress preview did not become ready. " +
-                "PreviewBody=${readTaggedTextOrNull(workspaceSettingsResetProgressPreviewBodyTag)} " +
-                "PreviewError=${workspaceSettingsResetProgressErrorTextOrNull()} " +
-                "CloudSettings=${currentCloudSettingsSummary()} " +
-                "CurrentWorkspace=${currentWorkspaceSummaryOrNull()}",
-            error
-        )
-    }
-}
-
-private fun LiveSmokeContext.readTaggedText(tag: String, label: String): String {
-    waitForTagToExist(
-        tag = tag,
-        timeoutMillis = externalUiTimeoutMillis,
-        context = "while reading $label"
-    )
-    val text = nodeSummary(composeRule.onNodeWithTag(tag).fetchSemanticsNode()).trim()
-    if (text.isBlank()) {
-        throw AssertionError("Tagged text for $label was blank.")
-    }
-
-    return text
-}
-
-private fun LiveSmokeContext.readTaggedTextOrNull(tag: String): String? {
-    return composeRule.onAllNodesWithTag(tag).fetchSemanticsNodes()
-        .singleOrNull()
-        ?.let(::nodeSummary)
-        ?.trim()
-        ?.takeIf { text -> text.isNotBlank() }
-}
-
-private fun LiveSmokeContext.workspaceSettingsResetProgressErrorTextOrNull(): String? {
-    return composeRule.onAllNodesWithTag(workspaceSettingsResetProgressErrorTag).fetchSemanticsNodes()
-        .singleOrNull()
-        ?.let(::nodeSummary)
-        ?.trim()
-        ?.takeIf { text -> text.isNotBlank() }
-}
-
-private fun extractResetProgressCount(previewBodyText: String): Int {
-    val match = Regex("""reset progress for (\d+) card""").find(previewBodyText)
-        ?: throw AssertionError("Reset progress preview body did not contain a card count: '$previewBodyText'")
-    return match.groupValues[1].toInt()
-}
-
-internal fun LiveSmokeContext.dismissAiConsentIfNeeded() {
-    if (composeRule.onAllNodesWithText("Before you use AI").fetchSemanticsNodes().isNotEmpty()) {
-        clickText(text = "OK", substring = false)
-    }
-}
-
-internal fun LiveSmokeContext.updateCardText(fieldTitle: String, value: String) {
-    clickText(text = fieldTitle, substring = false)
-    composeRule.onAllNodes(hasSetTextAction())[0].performTextReplacement(value)
-    tapBackIcon()
-}
-
-internal fun LiveSmokeContext.waitForSelectedWorkspaceSummary(context: String, timeoutMillis: Long) {
-    try {
-        scrollCurrentWorkspaceListToSelectedWorkspace()
-        waitUntilWithMitigation(
-            timeoutMillis = timeoutMillis,
-            context = "while waiting for current workspace selection $context"
-        ) {
-            selectedWorkspaceSummaryOrNull() != null
-        }
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Current workspace selection did not settle $context. " +
-                "Visible linked workspaces=${captureVisibleWorkspaceRows(rowTag = currentWorkspaceExistingRowTag)} " +
-                "Current workspace name=${currentWorkspaceNameOrNull()}",
-            error
-        )
-    }
-}
-
-internal fun LiveSmokeContext.waitForCurrentWorkspaceScreenToSettle() {
-    try {
-        waitUntilWithMitigation(
-            timeoutMillis = externalUiTimeoutMillis,
-            context = "while waiting for the Current Workspace screen to settle"
-        ) {
-            val visibleError: String? = currentWorkspaceVisibleErrorMessageOrNull()
-            if (visibleError != null) {
-                throw AssertionError("Current Workspace settled with an error: $visibleError")
-            }
-
-            val isLoading: Boolean = composeRule.onAllNodesWithTag(currentWorkspaceLoadingStateTag)
-                .fetchSemanticsNodes()
-                .isNotEmpty()
-            isLoading.not() && (
-                composeRule.onAllNodesWithTag(currentWorkspaceCreateButtonTag)
-                    .fetchSemanticsNodes()
-                    .isNotEmpty() ||
-                    composeRule.onAllNodesWithTag(currentWorkspaceReloadButtonTag)
-                        .fetchSemanticsNodes()
-                        .isNotEmpty()
-                )
-        }
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Current Workspace screen did not settle. " +
-                "Loading=${composeRule.onAllNodesWithTag(currentWorkspaceLoadingStateTag).fetchSemanticsNodes().isNotEmpty()} " +
-                "Error=${currentWorkspaceVisibleErrorMessageOrNull()} " +
-                "SelectedRow=${selectedWorkspaceSummaryOrNull()}",
-            error
-        )
-    }
-}
-
-private fun LiveSmokeContext.waitForSelectedWorkspaceSummaryToChange(
-    beforeSummary: String,
-    context: String,
-    timeoutMillis: Long
-) {
-    try {
-        waitUntilWithMitigation(
-            timeoutMillis = timeoutMillis,
-            context = "while waiting for current workspace selection to change $context"
-        ) {
-            runCatching {
-                scrollCurrentWorkspaceListToSelectedWorkspace()
-                selectedWorkspaceSummary(context = context) != beforeSummary
-            }.getOrDefault(false)
-        }
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Current workspace selection did not change $context. " +
-                "Before=$beforeSummary After=${selectedWorkspaceSummaryOrNull()} " +
-                "Current workspace name=${currentWorkspaceNameOrNull()} " +
-                "Error=${currentWorkspaceErrorMessageOrNull()}",
-            error
-        )
-    }
-}
-
-private fun LiveSmokeContext.waitForCurrentWorkspaceOperationToStart() {
-    try {
-        waitForTagToExist(
-            tag = currentWorkspaceOperationMessageTag,
-            timeoutMillis = internalUiTimeoutMillis,
-            context = "while waiting for current workspace operation to start"
-        )
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Current workspace operation did not start after tapping create. " +
-                "SelectedRow=${selectedWorkspaceSummaryOrNull()} " +
-                "Current workspace name=${currentWorkspaceNameOrNull()} " +
-                "Error=${currentWorkspaceErrorMessageOrNull()}",
-            error
-        )
-    }
-}
-
 private fun LiveSmokeContext.waitForCurrentWorkspaceOperationToFinish() {
     waitForCurrentWorkspaceOperationToLeaveSwitchingState()
     try {
@@ -907,156 +453,6 @@ private fun LiveSmokeContext.waitForWorkspaceOverviewReady(
                 "Error=${workspaceOverviewErrorMessageOrNull()}",
             error
         )
-    }
-}
-
-internal fun LiveSmokeContext.waitForCurrentWorkspaceName(expectedWorkspaceName: String) {
-    try {
-        waitUntilWithMitigation(
-            timeoutMillis = internalUiTimeoutMillis,
-            context = "while waiting for Current Workspace top card to update"
-        ) {
-            currentWorkspaceNameOrNull() == expectedWorkspaceName
-        }
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Current Workspace top card did not update after rename. " +
-                "TopCard=${currentWorkspaceNameOrNull()} " +
-                "SelectedRow=${selectedWorkspaceSummaryOrNull()}",
-            error
-        )
-    }
-}
-
-internal fun LiveSmokeContext.selectedWorkspaceSummary(context: String): String {
-    val selectedSummary: String? = selectedWorkspaceSummaryOrNull()
-    return requireNotNull(selectedSummary) {
-        "Current workspace selection was missing $context."
-    }
-}
-
-internal fun LiveSmokeContext.selectedWorkspaceSummaryOrNull(): String? {
-    val taggedSelection: String? = selectedWorkspaceSummaryFromCurrentSemanticsTree()
-    if (taggedSelection != null) {
-        return taggedSelection
-    }
-    scrollCurrentWorkspaceListToSelectedWorkspace()
-    return selectedWorkspaceSummaryFromCurrentSemanticsTree()
-}
-
-private fun LiveSmokeContext.selectedWorkspaceSummaryFromCurrentSemanticsTree(): String? {
-    return selectedWorkspaceSummaryFromSemanticsTree(useUnmergedTree = false)
-        ?: selectedWorkspaceSummaryFromSemanticsTree(useUnmergedTree = true)
-}
-
-private fun LiveSmokeContext.selectedWorkspaceSummaryFromSemanticsTree(
-    useUnmergedTree: Boolean
-): String? {
-    return composeRule.onAllNodesWithTag(
-        testTag = currentWorkspaceSelectedSummaryTag,
-        useUnmergedTree = useUnmergedTree
-    )
-        .fetchSemanticsNodes()
-        .singleOrNull()
-        ?.let(::nodeSummaryIncludingDescendants)
-        ?.trim()
-        ?.takeIf { summary -> summary.isNotBlank() }
-}
-
-internal fun LiveSmokeContext.workspaceOverviewNameFieldValueOrNull(): String? {
-    return composeRule.onAllNodesWithTag(workspaceOverviewNameFieldTag)
-        .fetchSemanticsNodes()
-        .singleOrNull()
-        ?.config
-        ?.getOrNull(SemanticsProperties.EditableText)
-        ?.text
-}
-
-internal fun LiveSmokeContext.workspaceOverviewErrorMessageOrNull(): String? {
-    return composeRule.onAllNodesWithTag(workspaceOverviewErrorMessageTag)
-        .fetchSemanticsNodes()
-        .singleOrNull()
-        ?.let(::nodeSummary)
-}
-
-internal fun LiveSmokeContext.currentWorkspaceNameOrNull(): String? {
-    scrollCurrentWorkspaceListToTopCard()
-    return composeRule.onAllNodesWithTag(currentWorkspaceNameTag)
-        .fetchSemanticsNodes()
-        .singleOrNull()
-        ?.let(::nodeSummary)
-}
-
-internal fun LiveSmokeContext.currentWorkspaceErrorMessageOrNull(): String? {
-    scrollCurrentWorkspaceListToTopCard()
-    return currentWorkspaceVisibleErrorMessageOrNull()
-}
-
-internal fun LiveSmokeContext.currentWorkspaceVisibleErrorMessageOrNull(): String? {
-    return composeRule.onAllNodesWithTag(currentWorkspaceErrorMessageTag)
-        .fetchSemanticsNodes()
-        .singleOrNull()
-        ?.let(::nodeSummary)
-}
-
-internal fun LiveSmokeContext.currentWorkspaceOperationMessageOrNull(): String? {
-    scrollCurrentWorkspaceListToTopCard()
-    return composeRule.onAllNodesWithTag(currentWorkspaceOperationMessageTag)
-        .fetchSemanticsNodes()
-        .singleOrNull()
-        ?.let(::nodeSummary)
-}
-
-private fun LiveSmokeContext.scrollCurrentWorkspaceListToSelectedWorkspace() {
-    if (composeRule.onAllNodesWithTag(currentWorkspaceListTag).fetchSemanticsNodes().isEmpty()) {
-        return
-    }
-    runCatching {
-        composeRule.onNodeWithTag(currentWorkspaceListTag).performScrollToNode(
-            matcher = hasTestTag(currentWorkspaceSelectedSummaryTag)
-        )
-    }
-}
-
-private fun LiveSmokeContext.scrollCurrentWorkspaceListToTopCard() {
-    if (composeRule.onAllNodesWithTag(currentWorkspaceListTag).fetchSemanticsNodes().isEmpty()) {
-        return
-    }
-    composeRule.onNodeWithTag(currentWorkspaceListTag).performScrollToNode(
-        matcher = hasTestTag(currentWorkspaceNameTag)
-    )
-}
-
-internal fun LiveSmokeContext.captureVisibleWorkspaceRows(rowTag: String): List<String> {
-    return composeRule.onAllNodesWithTag(rowTag)
-        .fetchSemanticsNodes()
-        .map(::nodeSummary)
-}
-
-internal fun LiveSmokeContext.currentCloudSettingsSummary(): String {
-    return runBlocking {
-        val appGraph = (composeRule.activity.application as FlashcardsApplication).appGraph
-        val cloudSettings = appGraph.cloudAccountRepository.observeCloudSettings().first()
-        "cloudState=${cloudSettings.cloudState} " +
-            "linkedUserId=${cloudSettings.linkedUserId} " +
-            "linkedWorkspaceId=${cloudSettings.linkedWorkspaceId} " +
-            "activeWorkspaceId=${cloudSettings.activeWorkspaceId} " +
-            "installationId=${cloudSettings.installationId}"
-    }
-}
-
-internal fun LiveSmokeContext.currentWorkspaceSummaryOrNull(): String? {
-    return runBlocking {
-        val appGraph = (composeRule.activity.application as FlashcardsApplication).appGraph
-        appGraph.workspaceRepository.observeWorkspace().first()?.let { workspace ->
-            "workspaceId=${workspace.workspaceId} name=${workspace.name}"
-        }
-    }
-}
-
-private fun LiveSmokeContext.currentActiveWorkspaceIdOrNull(): String? {
-    return runBlocking {
-        appGraph().cloudAccountRepository.observeCloudSettings().first().activeWorkspaceId
     }
 }
